+++ src/lib/data/queries.ts
@@ -17,6 +17,8 @@ export const getCodeListByUserQueryDocument = graphql(`
       account {
         uploader: address
       }
+      access_config_permission
+      access_config_addresses
     }
   }
 `);
@@ -29,6 +31,8 @@ export const getCodeListByIDsQueryDocument = graphql(`
       account {
         uploader: address
       }
+      access_config_permission
+      access_config_addresses
     }
   }
 `);
@@ -65,6 +69,52 @@ export const getInstantiateDetailByContractQueryDocument = graphql(`
       init_msg
       transaction {
         hash
+      }
+    }
+  }
+`);
+
+export const getExecuteTxsByContractAddress = graphql(`
+  query getExecuteTxsByContractAddress(
+    $contractAddress: String!
+    $offset: Int!
+    $pageSize: Int!
+  ) {
+    contract_transactions(
+      where: {
+        contract: { address: { _eq: $contractAddress } }
+        transaction: { is_execute: { _eq: true } }
+      }
+      order_by: { transaction: { block: { timestamp: desc } } }
+      limit: $pageSize
+      offset: $offset
+    ) {
+      transaction {
+        hash
+        messages
+        success
+        account {
+          address
+        }
+        block {
+          height
+          timestamp
+        }
+      }
+    }
+  }
+`);
+
+export const getExecuteTxsCountByContractAddress = graphql(`
+  query getExecuteTxsCountByContractAddress($contractAddress: String!) {
+    contract_transactions_aggregate(
+      where: {
+        contract: { address: { _eq: $contractAddress } }
+        transaction: { is_execute: { _eq: true } }
+      }
+    ) {
+      aggregate {
+        count
       }
     }
   }