+++ src/lib/services/codeService.ts
@@ -16,6 +16,8 @@ import type {
   ContractAddr,
   Option,
   ContractInfo,
+  InstantiatePermission,
+  PermissionAddresses,
 } from "lib/types";
 import { parseDateDefault, parseTxHashOpt, unwrap } from "lib/utils";
 
@@ -34,6 +36,10 @@ export const useCodeListByUserQuery = (
           id: code.id,
           contracts: code.contracts_aggregate.aggregate?.count ?? 0,
           uploader: code.account.uploader,
+          instantiatePermission:
+            code.access_config_permission as InstantiatePermission,
+          permissionAddresses:
+            code.access_config_addresses as PermissionAddresses,
         }))
       );
   }, [walletAddr]);
@@ -58,6 +64,10 @@ export const useCodeListByIDsQuery = (ids: Option<number[]>) => {
           id: code.id,
           uploader: code.account.uploader,
           contracts: code.instantiated,
+          instantiatePermission:
+            code.access_config_permission as InstantiatePermission,
+          permissionAddresses:
+            code.access_config_addresses as PermissionAddresses,
         }))
       );
   }, [ids]);
@@ -97,7 +107,8 @@ export const useCodeInfoByCodeId = (
                 ),
               }
             : undefined,
-          permissionAddresses: codes_by_pk.access_config_addresses,
+          permissionAddresses:
+            codes_by_pk.access_config_addresses as PermissionAddresses,
           instantiatePermission: codes_by_pk.access_config_permission,
         };
       });